feat: 아이템 기능 구현 (간단한 버전)

+++ Assets/02.Scripts/Bullet/Bullet.cs
@@ -5,7 +5,7 @@ public class Bullet : MonoBehaviour
 {
     // 목적 : 총알을 위로 움직이고 싶다. 
     public float MoveSpeed;
-    public float BulletDamage;
+    public int BulletDamage;
 
     private void Update()
     {
