+++ Assets/02.Scripts/Enemy/Enemy.cs
@@ -6,9 +6,9 @@ public abstract class Enemy : MonoBehaviour
 {
     // privvate로 원천 차단을 하자. 
     // 유니티가 수정할 수 있는 필드 - SerializeField
-    [SerializeField] private float _health = 100f;
+    [SerializeField] private int _health = 100;
     [SerializeField] protected float _moveSpeed = 5f;
-    [SerializeField] protected float _enemyDamage = 30f;
+    [SerializeField] protected int _enemyDamage = 30;
 
     protected virtual void Update()
     {
@@ -36,7 +36,7 @@ public abstract class Enemy : MonoBehaviour
         Destroy(gameObject);
     }
 
-    public void TakeDamage(float damage)
+    public void TakeDamage(int damage)
     {
         _health -= damage;
         if (_health <= 0)
